[twins] reorganize files
Closes: #6129
