+++ consensus/src/network_tests.rs
@@ -432,10 +432,7 @@ impl DropConfig {
     }
 
     pub fn drop_message_for(&mut self, src: &TwinId, dst: &TwinId) -> bool {
-        self.0
-            .entry(*src)
-            .or_insert_with(|| HashSet::new())
-            .insert(*dst)
+        self.0.entry(*src).or_insert_with(HashSet::new).insert(*dst)
     }
 
     pub fn split_network(
@@ -477,7 +474,7 @@ impl DropConfigRound {
         partition_first: &[TwinId],
         partition_second: &[TwinId],
     ) -> bool {
-        let config = self.0.entry(round).or_insert_with(|| DropConfig::default());
+        let config = self.0.entry(round).or_insert_with(DropConfig::default);
         config.split_network(partition_first, partition_second)
     }
 }
